+++ packages/web3-eth/test/integration/subscription_logs.test.ts
@@ -64,15 +64,13 @@ describeIf(isSocket)('subscription', () => {
 
 	beforeEach(async () => {
 		tempAcc = await createTempAccount();
-	});
-	beforeAll(() => {
 		clientUrl = getSystemTestProviderUrl();
 		provider = isWs ? new WebSocketProvider(clientUrl) : new IpcProvider(clientUrl);
 		contract = new Contract(BasicAbi, undefined, {
 			provider,
 		});
 	});
-	afterAll(async () => {
+	afterEach(async () => {
 		provider.disconnect();
 		await closeOpenConnection(web3Eth);
 	});