+++ packages/web3-eth/test/unit/rpc_method_wrappers/get_transaction_receipt.test.ts
@@ -37,7 +37,7 @@ describe('getTransactionReceipt', () => {
 		async (_, inputParameters) => {
 			const [inputTransactionHash] = inputParameters;
 			const inputTransactionHashFormatted = format(
-				{ eth: 'bytes32' },
+				{ format: 'bytes32' },
 				inputTransactionHash,
 				DEFAULT_RETURN_FORMAT,
 			);